requirements.txt
matplotlib.pyplot
numpy
requests
pandas
flask 
werkzeug.security
click

